+++ .gitignore
@@ -1,0 +1,24 @@
+
+__pycache__/
+*.py[cod]
+*.pyo
+
+venv/
+.env
+*.env
+
+.vscode/
+
+.DS_Store
+Thumbs.db
+
+instance/
+*.log
+
+Pipfile.lock
+poetry.lock
+
+.pytest_cache/
+
+*.sqlite3
+*.db
