--- Backend-feedback/.gitignore
@@ -1,3 +1,0 @@
-__pycache__/
-venv/
-.env
