--- Backend-feedback/app/schemas.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-class FeedbackCreate(BaseModel):
-    role: str
-    trip_info: str
-    message: str
