+++ server/app/__init__.py
@@ -1,1 +1,1 @@
-ECHO is on.
+# ECHO is on.
