+++ server/config.py
@@ -1,0 +1,10 @@
+# config.py
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+class Config:
+    SECRET_KEY = os.getenv("SECRET_KEY", "fallback_secret_key")
+    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL", "sqlite:///feedback.db")
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
