+++ server/feedback/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+feedback_bp = Blueprint("feedback", __name__)
+
+from . import routes
