+++ server/feedback/models.py
@@ -1,9 +1,9 @@
-from . import db
 from datetime import datetime
+from . import db
 
 class Feedback(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    role = db.Column(db.String(50))
+    role = db.Column(db.String(50), nullable=False)
     trip_info = db.Column(db.String(255))
-    message = db.Column(db.Text)
+    message = db.Column(db.Text, nullable=False)
     submitted_at = db.Column(db.DateTime, default=datetime.utcnow)
